requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
httpretty==1.0.2
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2-binary==2.8.5
pycparser==2.20
pytz==2020.1
redis==3.5.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
